requirements.txt
Flask==1.0.2
Jinja2==2.10
Werkzeug==0.14.1
tqdm==4.28.1
pandas==0.23.4
numpy==1.15.4
Keras==2.2.4
Pillow==5.3.0
scikit_learn==0.20.2
